+++ src/atoms/helper/storage_atoms.ts
@@ -1,15 +1,13 @@
 import { atomWithStorage, createJSONStorage } from "../../deps.ts";
 
 const gmStorage = {
-  getItem: <T>(key: string, initialValue: T) => {
-    return GM_getValue(key, initialValue);
-  },
-  setItem: <T>(key: string, value: T) => GM_setValue(key, value),
+  getItem: GM_getValue,
+  setItem: GM_setValue,
   removeItem: (key: string) => GM_deleteValue(key),
 };
 
 export function gmValueAtom<T>(key: string, defaultValue: T) {
-  return atomWithStorage<T>(key, defaultValue, gmStorage);
+  return atomWithStorage<T>(key, GM_getValue(key, defaultValue), gmStorage);
 }
 
 const jsonSessionStorage = createJSONStorage(() => sessionStorage);